得到前端代码

+++ web/js/common.js
@@ -1,4 +1,3 @@
-ENV=''
 //  ajaxSetup
 // ============================================================
 $.ajaxSetup({
@@ -6,15 +5,42 @@ $.ajaxSetup({
 });
 
 var editor;
+var role = 1; //角色标记
+var menu;
+var menuStr;
+var menuUrl;
+var itemType;
 
 $(function() {
+
+    //  处理角色
+    // ============================================================
+    var user= $.cookie("role");
+    switch(user){
+        case('USER'):
+            $('#pre_login').hide();
+            $('#after_login').show();role=0;break;
+        case('MANAGER'):
+        case('VIP'):
+            $('#pre_login').hide();
+            $('#after_login').show();role=1;break;
+        default:
+            role=0;
+    }
+    if (role == 0) {
+        $('#edit-area').hide();
+    } else {
+        $('#edit-area').show();
+    }
+
+    //  设置面板
+    // ============================================================
+    setPanel();
+
     //  编辑器初始化
     // ============================================================
     editor = UE.getEditor('editor', {
-        serverUrl:  'news/add/ue',
-        zIndex: 1,
-        elementPathEnabled: false,
-        wordCount: false
+        serverUrl: menuUrl + '/add/ue',
     });
 
     //  添加内容
@@ -24,15 +50,16 @@ $(function() {
             var data = {
                 title: $("#name").val(),
                 content: editor.getContent(),
-                newsId: $('#name').attr('editId'),
+                id: $('#name').attr('editId'),
             };
-            if(checkContent(data) == true) modifyContent(data);
+            if (checkContent(data) == true) modifyContent(data);
         } else {
             var data = {
                 title: $("#name").val(),
                 content: editor.getContent(),
+                type: itemType,
             };
-            if(checkContent(data) == true) addContent(data);
+            if (checkContent(data) == true) addContent(data);
         }
     })
 
@@ -42,11 +69,44 @@ $(function() {
 
 })
 
+//  设置面版
+// ============================================================
+function setPanel() {
+    var request = getRequest();
+    if (request['menu'] == undefined) return;
+    menuStr = request['menu'];
+    menu = NAV[menuStr];
+    var menuList = menu['list'];
+    var type = request['type'];
+    var item;
+
+    // 设置左面板
+    $('#panel-header').html(menu.name);
+    $('#panel-list').empty();
+    menuUrl = menu.url;
+    for (var i = 0; i < menuList.length; i++) {
+        if (menu.hasUrl == 'true') {
+            $('#panel-list').append('<li class="list-group-item no-radius"><a href="' + menuList[i].url +'">' + menuList[i].title + '</a></li>');
+        } else {
+            var itemUrl = '/index.html?menu=' + menuStr + '&type=' + menuList[i].type;
+            $('#panel-list').append('<li class="list-group-item no-radius"><a href="' + itemUrl +'">' + menuList[i].title + '</a></li>');
+        }
+        if (menuList[i].type == type) {
+            item = menuList[i];
+        }
+    }
+
+    //设置右面板
+    $("#item-header").html(menu.name + ' - ' + item.title);
+    $("#item-title").html(item.title);
+    itemType = item.type;
+}
+
 //  添加内容
 // ============================================================
 function addContent(data) {
     $.ajax({
-            url: ENV + '/news/add/submit',
+            url: menuUrl + '/add/submit',
             method: 'post',
             data: data
         })
@@ -56,16 +116,24 @@ function addContent(data) {
         .done(function() {
             updateContentList()
             $("#name").val('');
-            editor.setContent('');
+            editor.execCommand('cleardoc');
+            reInitEditor();
             alertInfo('添加成功！');
         })
+}
+
+function reInitEditor() {
+    editor.destroy();
+    editor = UE.getEditor('editor', {
+        serverUrl: menuUrl + '/add/ue',
+    });
 }
 
 //  修改内容
 // ============================================================
 function modifyContent(data) {
     $.ajax({
-            url: ENV + '/news/update/submit',
+            url: menuUrl + '/update/submit',
             method: 'post',
             data: data
         })
@@ -74,8 +142,9 @@ function modifyContent(data) {
         })
         .done(function() {
             updateContentList()
-            editor.setContent('');
+            editor.execCommand('cleardoc');
             $("#name").val('');
+            reInitEditor();
             $("#name").attr('edit', false);
             $("#name").attr('editId', -1);
             alertInfo('修改成功！');
@@ -99,21 +168,22 @@ function checkContent(data) {
 function getContentList() {
     $("#contentPage").pagination({
         pageIndex: 0,
-        pageSize:10,
+        pageSize: 10,
         total: 100,
         debug: false,
         showInfo: true,
         showJump: false,
-        infoFormat: '{start} ~ {end}条，共{total}条',
+        infoFormat: '共{total}条',
         showPageSizes: false,
         pageElementSort: ['$page', '$size', '$jump', '$info'],
         remote: {
-            url: ENV + '/news/list',
+            url: menuUrl + '/list',
             pageParams: function(data) {
                 return {
                     size: data.pageSize,
                     page: data.pageIndex,
-                    sort: encodeURI("id,desc")
+                    type: itemType,
+                    //sort: encodeURI("id,desc")
                 };
             },
             totalName: 'page.totalElements',
@@ -140,13 +210,33 @@ $("#contentPage").on("pageClicked", function(event, data) {
 //  更新表格
 // ============================================================
 function updateContentTable(content) {
+    $("#contentHead").empty();
+    if (role == 0) {
+        var line = $('<tr></tr>');
+        line.append('<th class="col-md-6">标题</th>');
+        line.append('<th class="col-md-2">时间</th>');
+        line.append('<th class="col-md-1">发布人</th>');
+        $('#contentHead').append(line);
+    } else {
+        var line = $('<tr></tr>');
+        line.append('<th class="col-md-4">标题</th>');
+        line.append('<th class="col-md-2">时间</th>');
+        line.append('<th class="col-md-1">发布人</th>');
+        line.append('<th class="col-md-1">修改</th>');
+        line.append('<th class="col-md-1">删除</th>');
+        $('#contentHead').append(line);
+    }
+
     $("#contentTable").empty();
     for (var i = 0; i < content.length; i++) {
         var line = $('<tr itemId=' + content[i].id + '></tr>');
-        line.append('<td><a href="/view.html?menu=news&id=' + content[i].id + '" target="_blank" title=' + content[i].title + ' >' + sliceTitle(content[i].title, 20) + '</a></td>');
-        line.append('<td>' + transUTC(content[i].publishedAt) + '</td>');
-        line.append('<td><a href="#" onclick="editItem(this)">修改</a></td>');
-        line.append('<td><a href="#" onclick="deleteItem(this)">删除</a></td>');
+        line.append('<td><a href="/view.html?menu=' + menuStr + '&type=' + itemType + '&id=' + content[i].id + '" target="_blank" title=' + content[i].title + ' >' + sliceTitle(content[i].title, 20) + '</a></td>');
+        line.append('<td>' + transUTC(content[i].modifiedAt) + '</td>');
+        line.append('<td>' + content[i].author.username + '</td>');
+        if (role == 1) {
+            line.append('<td><a href="#" onclick="editItem(this)">修改</a></td>');
+            line.append('<td><a href="#" onclick="deleteItem(this)">删除</a></td>');
+        }
         $('#contentTable').append(line);
     }
 }
@@ -156,10 +246,10 @@ function updateContentTable(content) {
 function deleteItem(item) {
     var id = $(item).parent().parent().attr('itemId');
     $.ajax({
-            url: ENV + '/news/delete',
+            url: menuUrl + '/delete',
             method: 'post',
             data: {
-                newsId: id
+                id: id
             },
         })
         .fail(function(jqXHR, textStatus) {
@@ -167,6 +257,12 @@ function deleteItem(item) {
         })
         .done(function(data) {
             updateContentList();
+            if ($("#name").attr('edit') == 'true' && $("#name").attr('editId') == id) {
+                editor.execCommand('cleardoc');
+                $("#name").val('');
+                $("#name").attr('edit', false);
+                $("#name").attr('editId', -1);
+            }
             alertInfo('删除成功!');
         })
 }
@@ -178,25 +274,22 @@ function editItem(item) {
     //  编辑器初始化
     // ============================================================
     editor = UE.getEditor('editor', {
-        serverUrl: ENV + '/news/update/ue?newsId=' + id,
-        zIndex: 1,
-        elementPathEnabled: false,
-        wordCount: false
+        serverUrl: menuUrl + '/update/ue?id=' + id,
     });
 
     $.ajax({
-            url: ENV + '/news/get',
+            url: menuUrl + '/get',
             method: 'get',
             data: {
-                newsId: id
+                id: id
             },
         })
         .fail(function(jqXHR, textStatus) {
             alertWarning('获取失败!');
         })
         .done(function(data) {
-            $("#name").val(data.news.title);
-            editor.setContent(data.news.content);
+            $("#name").val(data.item.title);
+            editor.setContent(data.item.content);
             $("#name").attr('edit', true);
             $("#name").attr('editId', id);
         })
